chore: 명칭 변경으로 메인 화면에 넘어가지 않던 버그 수정
