+++ lib/Screen/PracticeNote/PracticeProblemSelectionScreen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:ono/Model/PracticeNoteModel.dart';
-import 'package:ono/Screen/ProblemPractice/PracticeTitleWriteScreen.dart';
+import 'package:ono/Screen/PracticeNote/PracticeTitleWriteScreen.dart';
 import 'package:provider/provider.dart';
 import '../../GlobalModule/Text/StandardText.dart';
 import '../../GlobalModule/Theme/ThemeHandler.dart';
