+++ lib/Screen/ProblemRegister/ProblemRegisterScreenWithTemplate.dart
@@ -6,24 +6,27 @@ import 'package:provider/provider.dart';
 import '../../GlobalModule/Text/StandardText.dart';
 import '../../GlobalModule/Theme/ThemeHandler.dart';
 import '../../Provider/FoldersProvider.dart';
-import 'ProblemRegisterTemplateV2.dart';
 
-class ProblemRegisterScreenV2 extends StatefulWidget {
-  final ProblemModel? problemModel;
+class ProblemRegisterScreen extends StatefulWidget {
+  final ProblemModel problemModel;
   final bool isEditMode;
+  final Map<String, dynamic>? colorPickerResult;
+  final List<List<double>>? coordinatePickerResult;
 
-  const ProblemRegisterScreenV2({
+  const ProblemRegisterScreen({
     super.key,
     required this.problemModel,
     required this.isEditMode,
+    required this.colorPickerResult,
+    required this.coordinatePickerResult,
   });
 
   @override
-  _ProblemRegisterScreenStateV2 createState() =>
-      _ProblemRegisterScreenStateV2();
+  _ProblemRegisterScreenState createState() =>
+      _ProblemRegisterScreenState();
 }
 
-class _ProblemRegisterScreenStateV2 extends State<ProblemRegisterScreenV2> {
+class _ProblemRegisterScreenState extends State<ProblemRegisterScreen> {
   @override
   void initState() {
     super.initState();
@@ -34,16 +37,27 @@ class _ProblemRegisterScreenStateV2 extends State<ProblemRegisterScreenV2> {
     Widget templateWidget;
     final themeProvider = Provider.of<ThemeHandler>(context);
 
-    templateWidget = ProblemRegisterTemplateV2(
+    templateWidget = ProblemRegisterTemplate(
         problemModel: widget.problemModel,
+        colorPickerResult: widget.colorPickerResult,
+        coordinatePickerResult: widget.coordinatePickerResult,
         isEditMode: widget.isEditMode,
-    );
+        templateType: widget.problemModel.templateType!);
 
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
         centerTitle: true,
         backgroundColor: Colors.white,
+        leading: IconButton(
+          icon: Icon(Icons.arrow_back, color: themeProvider.primaryColor),
+          onPressed: () {
+            if (!widget.isEditMode) {
+              _deleteProblem();
+            }
+            Navigator.pop(context);
+          },
+        ),
         title: StandardText(
           text: widget.isEditMode ? '오답노트 수정' : '오답노트 작성',
           color: themeProvider.primaryColor,
@@ -56,5 +70,10 @@ class _ProblemRegisterScreenStateV2 extends State<ProblemRegisterScreenV2> {
         child: templateWidget,
       ),
     );
+  }
+
+  Future<void> _deleteProblem() async {
+    await Provider.of<FoldersProvider>(context, listen: false)
+        .deleteProblems([widget.problemModel.problemId]);
   }
 }
