+++ lib/main.dart
@@ -11,14 +11,14 @@ import 'package:ono/GlobalModule/Theme/ThemeHandler.dart';
 import 'package:ono/Provider/FoldersProvider.dart';
 import 'package:ono/Provider/PracticeNoteProvider.dart';
 import 'package:ono/Provider/ScreenIndexProvider.dart';
+import 'package:ono/Screen/ProblemRegister/ProblemRegisterScreenWithTemplate.dart';
 import 'package:ono/Screen/ProblemRegister/ProblemRegisterScreen.dart';
-import 'package:ono/Screen/ProblemRegister/ProblemRegisterScreenV2.dart';
 import 'package:ono/Screen/SplashScreen.dart';
 import 'package:provider/provider.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'Provider/UserProvider.dart';
-import 'Screen/ProblemManagement/DirectoryScreen.dart';
-import 'Screen/ProblemPractice/PracticeThumbnailScreen.dart';
+import 'Screen/Folder/DirectoryScreen.dart';
+import 'Screen/PracticeNote/PracticeThumbnailScreen.dart';
 import 'Screen/ProblemRegister/TemplateSelectionScreen.dart';
 import 'Screen/SettingScreen.dart';
 import 'firebase_options.dart';
